+++ src/kingdom/chat-buttons.ts
@@ -24,6 +24,7 @@ export const kingdomChatButtons: KingdomChatButton[] = [
         callback: async (game: Game, actor: Actor, event: Event): Promise<void> => {
             event.preventDefault();
             const update = gainFame(getKingdom(actor), 1);
+            await ChatMessage.create({content: 'Gaining 1 Fame'});
             await saveKingdom(actor, update);
         },
     },