+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Added
+
+* Also display building from ruin cost when building structures
+
+### Fixed
+
+* When clicking Gain 1 Fame button on a critical success, a message is posted to chat as usual
+
 ## [0.43.0] - 2024-03-16
 
 ### Fixed